requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
packaging==25.0
passlib==1.7.4
pefile==2024.8.26
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-dotenv==1.2.1
python-jose==3.5.0
pywin32-ctypes==0.2.3
rsa==4.9.1
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid==1.30
uvicorn==0.39.0
zipp==3.23.0
